changing pin code and goto main activity on restart

+++ app/src/main/java/com/gasanovmagomed/cleversafe/ui/notifications/NotificationsFragment.kt
@@ -1,5 +1,6 @@
 package com.gasanovmagomed.cleversafe.ui.notifications
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
@@ -16,7 +17,6 @@ import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import com.gasanovmagomed.cleversafe.MainActivity
 import com.gasanovmagomed.cleversafe.R
-import com.gasanovmagomed.cleversafe.StorageActivity
 import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
 
@@ -25,9 +25,9 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
     private lateinit var auth: FirebaseAuth
     private lateinit var emailEntry: TextView
     private lateinit var email: String
+    private lateinit var prefs: SharedPreferences
     private lateinit var userPin: String
     private val savedUserPin = "usersPin"
-    private lateinit var prefs: SharedPreferences
 
     override fun onCreateView(
             inflater: LayoutInflater,
@@ -40,8 +40,8 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
         val changePinBtn = root.findViewById<Button>(R.id.changePin)
         val deleteUserBtn = root.findViewById<Button>(R.id.deleteAccount)
         emailEntry = root.findViewById(R.id.emailEntry)
-        userPin = ""
-        prefs = requireActivity().getPreferences(Context.MODE_PRIVATE)
+        prefs = requireActivity().getSharedPreferences("settings", Context.MODE_PRIVATE)
+        userPin = prefs.getString(savedUserPin, null).toString()
         changeMailBtn.setOnClickListener(this)
         changePassBtn.setOnClickListener(this)
         changePinBtn.setOnClickListener(this)
@@ -51,15 +51,66 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
         return root
     }
 
+    override fun onPause() {
+        super.onPause()
+
+        val editor = prefs.edit()
+        editor.clear().apply()
+        editor.putString(savedUserPin, userPin).apply()
+    }
+
+    override fun onResume() {
+        super.onResume()
+        if(prefs.contains(savedUserPin)) userPin = prefs.getString(savedUserPin, null).toString()
+    }
+
     override fun onClick(v: View?) {
-        when(setButtonIndex(v!!.id)){
-            1 -> changePin()
-            2 -> changePassword()
-            3 -> changeEmail()
-            4 -> deleteAccount()
+        val user = auth.currentUser!!
+
+        val builder = activity?.let { AlertDialog.Builder(it) }
+        val inflater: LayoutInflater = layoutInflater
+        val dialogLayout = inflater.inflate(R.layout.confirm_account_window, null)
+        val confirmedEmail: EditText = dialogLayout.findViewById((R.id.confirmedEmail))
+        val confirmedPassword: EditText = dialogLayout.findViewById((R.id.confirmedPassword))
+
+        with(builder) {
+            this?.setTitle("Confirm your account")
+            this?.setMessage("write here your password and email to confirm your account")
+            this?.setPositiveButton("Change") { dialog, which ->
+                val confirmedEmailText = confirmedEmail.text.toString()
+                val confirmedPasswordText = confirmedPassword.text.toString()
+
+                if (confirmedEmailText.isEmpty()) {
+                    confirmedEmail.error = "Поле 'email' обязательно для заполнения"
+                    return@setPositiveButton
+                }
+
+                if (confirmedPasswordText.isEmpty()) {
+                    confirmedPassword.error = "Поле 'password' обязательно для заполнения"
+                    return@setPositiveButton
+                }
+
+                val credential = EmailAuthProvider
+                    .getCredential(confirmedEmailText, confirmedPasswordText)
+                    user.reauthenticate(credential)
+                        .addOnCompleteListener {
+                            when(setButtonIndex(v!!.id)){
+                                1 -> changePin()
+                                2 -> changePassword()
+                                3 -> changeEmail()
+                                4 -> deleteAccount()
+                            }
+                        }
+            }
+            this?.setNegativeButton("Cancel") { dialog, which ->
+                showCancelMsg()
+            }
+            this?.setView(dialogLayout)
+            this?.show()
         }
     }
 
+    @SuppressLint("CommitPrefEdits")
     private fun changePin() {
         val builder = activity?.let { AlertDialog.Builder(it) }
         val inflater: LayoutInflater = layoutInflater
@@ -70,23 +121,16 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
             this?.setTitle("Change your pin code")
             this?.setMessage("write here your new pin code to change it")
             this?.setPositiveButton("Change") { dialog, which ->
-                val user = auth.currentUser
                 val newPinText = newPin.text.toString()
 
                 if (newPinText.isEmpty()) {
                     newPin.error = "Поле 'pin' обязательно для заполнения"
                     return@setPositiveButton
                 }
-
-                user!!.updatePassword(newPinText)
-                        .addOnCompleteListener { task ->
-                            if (task.isSuccessful) {
-                                userPin = newPinText
-                                val editor = prefs.edit()
-                                editor.putString(savedUserPin, userPin)
-                            }
-                        }
-
+                Log.d("check pin", userPin)
+                userPin = newPinText
+                Toast.makeText(activity, "Your pin code successfully changed", Toast.LENGTH_SHORT).show()
+                Log.d("check pin", userPin)
             }
             this?.setNegativeButton("Cancel") { dialog, which ->
                 showCancelMsg()
@@ -115,11 +159,11 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
                 }
 
                 user!!.updatePassword(newPasswordText)
-                        .addOnCompleteListener { task ->
-                            if (task.isSuccessful) {
-                                Log.d("Changing password", "User password updated.")
-                            }
+                    .addOnCompleteListener { task ->
+                        if (task.isSuccessful) {
+                            Log.d("Changing password", "User password updated.")
                         }
+                    }
 
             }
             this?.setNegativeButton("Cancel"){ dialog, which ->
@@ -135,7 +179,6 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
         val inflater: LayoutInflater = layoutInflater
         val dialogLayout = inflater.inflate(R.layout.change_email_window, null)
         val newEmail: EditText = dialogLayout.findViewById((R.id.newEmail))
-
         with(builder){
             this?.setTitle("Change your email address")
             this?.setMessage("write here your new email address to change it")
@@ -149,11 +192,11 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
                 }
 
                 user!!.updateEmail(newEmailText)
-                        .addOnCompleteListener { task ->
-                            if (task.isSuccessful) {
-                                Log.d("Update", "User email address updated.")
-                            }
+                    .addOnCompleteListener { task ->
+                        if (task.isSuccessful) {
+                            Log.d("Update", "User email address updated.")
                         }
+                    }
 
             }
             this?.setNegativeButton("Cancel"){ dialog, which ->
@@ -168,17 +211,16 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
         val user = auth.currentUser!!
 
         user.delete()
-                .addOnCompleteListener { task ->
-                    if (task.isSuccessful) {
-                        val editor = prefs.edit()
-                        editor.remove(savedUserPin);
-                        editor.apply()
-                        val intent = Intent(activity, MainActivity::class.java)
-                        startActivity(intent)
-                    } else{
-                        Toast.makeText(activity, "Somthing was wrong", Toast.LENGTH_LONG).show()
-                    }
+            .addOnCompleteListener { task ->
+                if (task.isSuccessful) {
+                    val editor = prefs.edit()
+                    editor.clear().apply()
+                    val intent = Intent(activity, MainActivity::class.java)
+                    startActivity(intent)
+                } else{
+                    Toast.makeText(activity, "Somthing was wrong", Toast.LENGTH_LONG).show()
                 }
+            }
     }
 
     private fun displayUsersInformation() {
@@ -201,47 +243,6 @@ class NotificationsFragment : Fragment(), View.OnClickListener {
             R.id.deleteAccount -> index = 4
         }
         return index
-    }
-
-    private fun confirmAccount() {
-        val user = auth.currentUser!!
-
-        val builder = activity?.let { AlertDialog.Builder(it) }
-        val inflater: LayoutInflater = layoutInflater
-        val dialogLayout = inflater.inflate(R.layout.confirm_account_window, null)
-        val confirmedEmail: EditText = dialogLayout.findViewById((R.id.confirmedEmail))
-        val confirmedPassword: EditText = dialogLayout.findViewById((R.id.confirmedPassword))
-
-        with(builder){
-            this?.setTitle("Confirm your account")
-            this?.setMessage("write here your password and email to confirm your account")
-            this?.setPositiveButton("Change"){ dialog, which ->
-                val confirmedEmailText = confirmedEmail.text.toString()
-                val confirmedPasswordText = confirmedPassword.text.toString()
-
-                if(confirmedEmailText.isEmpty()){
-                    confirmedEmail.error = "Поле 'email' обязательно для заполнения"
-                    return@setPositiveButton
-                }
-
-                if(confirmedPasswordText.isEmpty()){
-                    confirmedPassword.error = "Поле 'password' обязательно для заполнения"
-                    return@setPositiveButton
-                }
-
-                val credential = EmailAuthProvider
-                        .getCredential(confirmedEmailText, confirmedPasswordText)
-                user.reauthenticate(credential)
-                        .addOnCompleteListener { Toast.makeText(activity, "Account has been confirmed", Toast.LENGTH_SHORT).show() }
-
-            }
-            this?.setNegativeButton("Cancel"){ dialog, which ->
-                showCancelMsg()
-            }
-            this?.setView(dialogLayout)
-            this?.show()
-        }
-
     }
 
 }
